Added resttemplate and connected with tokenService to check role from token

+++ DefectService/DefectService/src/main/java/com/microservice/DefectService/controller/VehicleController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -26,13 +27,16 @@ public class VehicleController {
 	
 	
 	@PostMapping("/add")
-    public ResponseEntity<String> addVehicle(@RequestBody CreateVehicleRequest request) {
-        try {
-            vehicleService.add(request);
-            return ResponseEntity.ok("Araç başarıyla eklendi.");
-        } catch (Exception ex) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ex.getMessage());
-        }
+    public ResponseEntity<String> addVehicle(@RequestBody CreateVehicleRequest request,
+    		@RequestHeader("Authorization") String authorizationHeader) {
+        
+		if (vehicleService.checkOperatorRole(authorizationHeader)) {
+        	 vehicleService.add(request);
+             return ResponseEntity.ok("Araç başarıyla eklendi.");
+		}
+		
+            return new ResponseEntity<>("Araç eklenemdi",HttpStatus.BAD_REQUEST);
+        
     }
 	
 	
@@ -40,15 +44,27 @@ public class VehicleController {
 	public ResponseEntity<Page<VehicleResponse>> getVehicles(@RequestParam int no,
 			@RequestParam int size,
 			@RequestParam String sortBy,
-			@RequestParam String sortDirection) {
-		Page<VehicleResponse> vehicles = vehicleService.getVehicles(no, size,sortBy,sortDirection);
-	 return new ResponseEntity<>(vehicles, HttpStatus.OK);
+			@RequestParam String sortDirection,
+			@RequestHeader("Authorization") String authorizationHeader) {
+		
+		if (vehicleService.checkTeamLeaderRole(authorizationHeader)) {
+			Page<VehicleResponse> vehicles = vehicleService.getVehicles(no, size,sortBy,sortDirection);
+			 return new ResponseEntity<>(vehicles, HttpStatus.OK);
+		}
+		return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
 	}
 	
-	 @DeleteMapping("/{vehicleId}")
-	    public String delete(@PathVariable String vehicleId) {
-		 vehicleService.delete(vehicleId);
-	        return "deleted";
+	
+	@DeleteMapping("/{vehicleId}")
+	    public String delete(@PathVariable String vehicleId,
+	    		@RequestHeader("Authorization") String authorizationHeader) {
+		 if (vehicleService.checkOperatorRole(authorizationHeader)) {
+			 vehicleService.delete(vehicleId);
+		        return "deleted";
+		}
+		 
+		 return "vehicle not deleted";
+		
 	    }
 
 }
